feat: MeetingConfirm에서 footer 분리
